Move all referral-related to referrals module
