+++ src/modules/projectAdministration/components/HmisParticipationsPage.tsx
@@ -3,7 +3,6 @@ import { Button, Paper } from '@mui/material';
 
 import { useCallback, useMemo } from 'react';
 
-import { useProjectDashboardContext } from '../../projects/components/ProjectDashboard';
 
 import { ColumnDef } from '@/components/elements/table/types';
 import PageTitle from '@/components/layout/PageTitle';
@@ -12,6 +11,7 @@ import { useViewEditRecordDialogs } from '@/modules/form/hooks/useViewEditRecord
 import HmisEnum from '@/modules/hmis/components/HmisEnum';
 import { HudRecordMetadataHistoryColumn } from '@/modules/hmis/components/HudRecordMetadata';
 import { parseAndFormatDateRange } from '@/modules/hmis/hmisUtil';
+import { useProjectDashboardContext } from '@/modules/projects/components/ProjectDashboard';
 import { cache } from '@/providers/apolloClient';
 import { HmisEnums } from '@/types/gqlEnums';
 import {
