+++ src/modules/projectAdministration/components/InventoriesPage.tsx
@@ -1,12 +1,12 @@
 import AddIcon from '@mui/icons-material/Add';
 import { Paper } from '@mui/material';
 
-import { useProjectDashboardContext } from '../../projects/components/ProjectDashboard';
 import InventoryTable from './InventoryTable';
 
 import ButtonLink from '@/components/elements/ButtonLink';
 import PageTitle from '@/components/layout/PageTitle';
 import { ProjectPermissionsFilter } from '@/modules/permissions/PermissionsFilters';
+import { useProjectDashboardContext } from '@/modules/projects/components/ProjectDashboard';
 import { ProjectDashboardRoutes } from '@/routes/routes';
 import { generateSafePath } from '@/utils/pathEncoding';
