+++ src/modules/projectAdministration/components/ProjectCocsPage.tsx
@@ -1,11 +1,11 @@
 import AddIcon from '@mui/icons-material/Add';
 import { Paper } from '@mui/material';
 
-import { useProjectDashboardContext } from '../../projects/components/ProjectDashboard';
 import ProjectCocTable from './ProjectCocTable';
 
 import ButtonLink from '@/components/elements/ButtonLink';
 import PageTitle from '@/components/layout/PageTitle';
+import { useProjectDashboardContext } from '@/modules/projects/components/ProjectDashboard';
 import { ProjectDashboardRoutes } from '@/routes/routes';
 import { generateSafePath } from '@/utils/pathEncoding';
