+++ src/modules/referrals/components/ProjectReferralPostingsTable.tsx
@@ -7,7 +7,7 @@ import {
   parseAndFormatDateTime,
 } from '@/modules/hmis/hmisUtil';
 import ReferralPostingStatusDisplay from '@/modules/referrals/components/ReferralPostingStatusDisplay';
-import { useReferralFilter } from '@/modules/referrals/components/useReferralFilter';
+import { useReferralFilter } from '@/modules/referrals/hooks/useReferralFilter';
 import { ProjectDashboardRoutes } from '@/routes/routes';
 import {
   GetProjectReferralPostingsDocument,
