+++ src/modules/referrals/components/ReferralPostingDetails.tsx
